Renaming and flattening in IDE
Added rename and flatten options to the IDE menu. The menu can be accessed by rightclicking in a .ql-file and hovering over the "Edit QL-file" option.

When renaming a variable, a prompt is shown to the user where the new name can be inserted.

When renaming, the selected location only needs to approximately match a location in the useDef rather than exactly.
If no matching location is found, an exception is now thrown.